重構和新增功能
在 `cxxwintoast.h` 中新增 `CXXToastResultCallback` 的 `typedef` 和相關註釋。
在 `cxxwritelocker.h` 中將析構函數改為虛擬函數。
在 `cxxcom.cpp` 中將函數名稱改為 `get`。
在 `cxxdir.cpp` 中新增目錄重命名函數。
在 `cxxfile.cpp` 中新增析構函數清理代碼，並更新路徑。
在 `cxxfile.cpp` 中對 `map` 和 `unmap` 函數增加互斥鎖。
在 `cxxfiledialog.cpp` 中重命名函數並移除重複代碼。
在 `cxxmutex.cpp` 中修改參數類型。
在 `cxxscopedhandle.cpp` 中新增獲取句柄函數。
在 `cxxstring.cpp` 和 `cxxvariant.cpp` 中更新註釋別名。

+++ CXXCore-master/src/cxxmutex.cpp
@@ -67,7 +67,7 @@ bool CXXMutex::tryLock()
 	return locked;
 }
 
-bool CXXMutex::tryLockFor(UINT duration)
+bool CXXMutex::tryLockFor(__int64 duration)
 {
 	std::chrono::system_clock::time_point start = std::chrono::system_clock::now();
 	std::chrono::system_clock::time_point end = start + std::chrono::milliseconds(duration);
